requirements.txt
pybit==5.10.1
pytrends==4.9.2
google-genai==1.16.1
numpy==2.2.5
yfinance==0.2.61
requests==2.32.3
tradingview-ta==3.3.0